moved add_move to battle_move_maker
