+++ src/battle/src/battle_move_maker.c
@@ -39,3 +39,23 @@ int build_moves(combatant_t *c)
     }
     return SUCCESS;
 }
+
+
+/* see battle_move_maker.h */
+int add_move(combatant_t *combatant, move_t *move) {
+    assert(combatant != NULL);
+    assert(move != NULL);
+
+    move_t *last_move = combatant->moves;
+    
+    if (combatant->moves == NULL){
+        combatant->moves = move;
+        return SUCCESS;
+    }
+
+    while (last_move->next != NULL) {
+        last_move = last_move->next;
+    }
+    last_move->next = move;
+    return SUCCESS;
+}
