+++ tests/battle/test_battle_move_maker.c
@@ -6,6 +6,7 @@
 #include "battle/battle_flow.h"
 #include "../include/battle/battle_move_maker.h"
 #include "playerclass/class.h"
+#include "battle/battle_default_objects.h"
 
 Test(class_moves, bard)
 {
@@ -105,4 +106,31 @@ Test(class_moves, knight)
 
     cr_assert_eq(ret_player->moves->defense, 0,
                  "add_class_move() didn't set defense");
+}
+
+Test(add_move, add_move_existing_list) {
+
+    move_t *old_move = get_random_default_move();
+    move_t *new_move = get_random_default_move();
+    battle_item_t *item = get_random_default_battle_item();
+    stat_t *cstats = get_random_stat();
+    combatant_t *player = combatant_new("TESTER", true, NULL, cstats, old_move, item, 0);
+
+    int res = add_move(player,new_move);
+    cr_assert_eq(player->moves->next,
+                 new_move,
+                 "add_move() did not add the move correctly");
+}
+
+Test(add_move, add_move_empty_list) {
+    
+    move_t *new_move = get_random_default_move();
+    battle_item_t *item = get_random_default_battle_item();
+    stat_t *cstats = get_random_stat();
+    combatant_t *player = combatant_new("TESTER",true,NULL,cstats,NULL,item, 0);
+
+    int res = add_move(player,new_move);
+    cr_assert_eq(player->moves,
+                 new_move,
+                 "add_move() did not add the move correctly");
 }
